@@ -1,26 +1,26 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import App from './App'
+import { createApp } from 'vue'
+import App from './App.vue'
 import router from './router'
 import axios from 'axios'
 import VueAxios from 'vue-axios'
-import {store} from './vuex/store'
+import Toast from 'vue-toastification'
+//import "vue-toastification/dist/index.css"
 
-Vue.use(VueAxios, axios)
-Vue.use(require('vue-moment'))
-Vue.config.productionTip = false
+const options = {
+    position: "bottom-right",
+    timeout: 60000,
+    closeOnClick: true,
+    pauseOnFocusLoss: false,
+    pauseOnHover: true,
+    draggable: true,
+    draggablePercent: 0.6,
+    showCloseButtonOnHover: false,
+    closeButton: false,
+    rtl: false,
+    icon: false
+};
 
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  created () {
-    if (this.$store.getters.isAuthenticated) {
-      this.$store.dispatch('userRequest')
-    }
-  },
-  router,
-  store,
-  components: { App },
-  template: '<App/>'
-})
+const app = createApp(App)
+app.use(VueAxios, axios)
+app.use(Toast, options)
+app.use(router).mount('#app')
